backend/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
uritemplate==4.1.1
